personal_care_chatbot/requirements.txt
requests
beautifulsoup4
pandas
psycopg2-binary
flask
openai
python-dotenv